requirements.txt
click==8.0.3
Flask==2.0.2
Flask-Cors==3.0.10
itsdangerous==2.0.1
Jinja2==3.0.3
MarkupSafe==2.0.1
numpy==1.22.0
Pillow==9.0.0
python-dotenv==0.19.2
six==1.16.0
torch
torchaudio==0.10.1
torchvision==0.11.2
typing_extensions==4.0.1
Werkzeug==2.0.2